Move getAppSettings to AbstractFunctionMojo
